Form and View changes, plus Requirements.txt update
A bug in the form and view functions of the user application had a
persistent bug which prevented the form's data from still persisting
after a reload and also displaying the error flagged by some particular
field, say the email field, for duplicates for example.

The requirements file has also been updated with newest versions of the
modules used.