Bug fixes and Code Refactor
Fixed a bug in settings.py and backends.py which prevented new
superusers from logging into the system.

gitignore now tracks user_applications directory but not its contents.

For majority of the HTML, most of the code was refactored and improved.
The base_auth.html file abstracted some of the repeated code in a number
of the auth HTML files so as to follow the DRY principle.

Made some errors in the send_approval_email function previously. Fixed
them.

The rest were just slight improvements (visually).